ai_service/requirements.txt
flask
flask-cors
textblob
python-dotenv
openai
